LayoutTests/media/modern-media-controls/time-control/time-control-expected.txt

Testing the TimeControl class.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS timeControl.element.localName is "div"
PASS timeControl.element.className is "time-control"

PASS timeControl.elapsedTimeLabel instanceof TimeLabel is true
PASS timeControl.scrubber instanceof Slider is true
PASS timeControl.remainingTimeLabel instanceof TimeLabel is true

timeControl.width = 500;
PASS timeControl.children.length is 3
PASS timeControl.children[0] is timeControl.elapsedTimeLabel
PASS timeControl.children[1] is timeControl.scrubber
PASS timeControl.children[2] is timeControl.remainingTimeLabel
PASS timeControl.elapsedTimeLabel.x is 0
PASS timeControl.scrubber.x is 32
PASS timeControl.scrubber.width is 430
PASS timeControl.remainingTimeLabel.x is 467

PASS successfullyParsed is true

TEST COMPLETE
